+++ packages/js-drive/lib/core/fetchTransactionFactory.js
@@ -1,0 +1,3 @@
+function fetchTransactionFactory() {
+  function fetchTransaction
+}
